% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NESTscore.R
\name{batch.table.freq}
\alias{batch.table.freq}
\title{batch.table.freq}
\usage{
batch.table.freq(x, batch_names)
}
\arguments{
\item{x}{factor vector: batch vector}

\item{batch_names}{character vector: batch names}
}
\description{
Internal helper function.
}
