% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NESTscore.R
\name{KLD}
\alias{KLD}
\title{KLD}
\usage{
KLD(P, Q)
}
\arguments{
\item{P}{probability distribution P}

\item{Q}{probability distribution Q}
}
\description{
Computes the Kullback-Leibler-Divergence for distributions P and Q.
}
