% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NESTscore.R
\name{NEST_search_k}
\alias{NEST_search_k}
\title{NEST_search_k}
\usage{
NEST_search_k(
  obj,
  group_by,
  k_search_vector = seq(2, 100, length.out = 50),
  color_plot_initial = T,
  k_initial = k_search_vector[round(length(k_search_vector)/2)],
  thresh_initial = NULL,
  reduction_in = "pca",
  ndims = 50,
  assay_in = Seurat::DefaultAssay(obj)
)
}
\arguments{
\item{obj}{Seurat object}

\item{group_by}{name of metadata-column with batch assignment}

\item{k_search_vector}{minimum and maximum number of nearest neighbors to be considered, maximal number has to be #cells -1.}

\item{color_plot_initial}{output plot is colored by cell-wise binary assignment based on \code{k_initial} and \code{thresh_initial}}

\item{k_initial}{number of nearest neighbors used for initial binary assignment for plotting}

\item{thresh_initial}{NEST-score threshold used for initial binary assignment for plotting. If \code{NULL} (default), it will be computed by \code{NEST_threshold_proposal()} with default parameters.}

\item{reduction_in}{name of dimensionality reduction to be considered for nearest neighbor search}

\item{ndims}{number of dimensions in dimensionality reduction to be considered for nearest neighbor search}

\item{assay_in}{name of seurat object assay}
}
\description{
Plots the NEST-Score for all cells across multiple choices of \code{k} nearest neighbors.
}
