% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NESTscore.R
\name{NESTscore}
\alias{NESTscore}
\title{NESTscore}
\usage{
NESTscore(
  obj,
  group_by,
  k_nn = 100,
  reduction_in = "pca",
  ndims = 50,
  assay_in = Seurat::DefaultAssay(obj),
  return_pairwise_eval = FALSE,
  show_heatmap = FALSE,
  heatmap_col = "darkred",
  name_suffix = ""
)
}
\arguments{
\item{obj}{Seurat object}

\item{group_by}{name of metadata-column with batch assignment}

\item{k_nn}{number of nearest neighbors to be considered. If set to "cellwise", a cell-wise k is computed based on a the connectivity of a fuzzy umap graph (computed with \code{uwot::umap} default parameters on the considered dimensionality reduction).}

\item{reduction_in}{name of dimensionality reduction to be considered for nearest neighbor search}

\item{ndims}{number of dimensions in dimensionality reduction to be considered for nearest neighbor search}

\item{assay_in}{name of seurat object assay}

\item{return_pairwise_eval}{return pairwise batch-to-batch evaluation as matrix}

\item{show_heatmap}{plots pairwise batch-to-batch matrix as heatmap (only considered if \code{return_pairwise_eval = TRUE})#' @param name_suffix additional suffix for new NEST-Score metadata column name}

\item{heatmap_col}{color for the optimal batch frequency in the pairwise batch-to-batch heatmap (only considered if \code{return_pairwise_eval = TRUE} and \code{show_heatmap = TRUE})}

\item{name_suffix}{additional suffix for new NEST-Score metadata column name}
}
\value{
Returns a list including the Seurat object \code{seuratobj} with cell-wise NEST-Score stored in an additional metadata column and the theoretical upper and lower bound of the NEST-Score \code{NESTscore_limits} for the considered batch assignment.
If \code{return_pairwise_eval = TRUE} the list also includes a pairwise batch-to-batch evaluation matrix \code{pairwise_matrix} and helpful plotting parameters \code{proposed_params_pheatmap} for using \code{pheatmap()}.
}
\description{
Computes the cell-wise NEST-Score for a given batch variable. For the NEST-Score the frequency vector of batches is computed for each cell considering their k-nearest neighborhood (knn).
This vector is divided by the global frequency vector of batches and scaled to add up to 1. To compare the resulting distribution vector against a uniform distribution, the negative Kullback-Leibler-Divergence is computed.
A high NEST-Score (close to 0) means, that the distribution of samples in the cell's knn matches the global distribution. A low NEST-Score (negative) means that a cell's knn is a less homogeneous, i.e. consists of fewer batches and/or different batch frequencies than the global distribution.
If all considered batches would mix randomly throughout the data, we would expect a NEST-Score close to 0 for all cells.
}
